Remove the vendor directory
The vendor directory ensures reproducible builds even if e.g. GitHub is
down or unreachable. However it is also dozens of MiBs that need to be
regularly updated. It is also no longer automatically used by `go`
unless in `-mod=vendor` mode.

My take on this is that anyone concerned about modules disappearing from
the ecosystem can run `go mod vendor` locally, and/or set up a module
proxy. As for reproducibility, go.sum has enough information to ensure
that builds are identical for a given Go version.